test(stella-core): split subagent/tests.rs into a directory by subject
subagent/tests.rs held 1426 lines against the strict 1500-line ceiling,
74 lines of headroom. It already carried section-comment dividers
matching the module doc's own "five contracts" framing, so split along
those seams into subagent/tests/{witness,tools_and_budget,
failure_and_events,seams,spend_and_cancellation}.rs, the same directory
pattern driver/tests/ already uses. The shared fakes (NoSleep,
ScriptedProvider, MixedTools, text_result/tool_call_result, drain,
finished) stay in subagent/tests.rs itself since every group calls
them, reached the same way driver/tests.rs's helpers already reach its
own submodules. Pure test move, no production change.

Closes #3709
